requirements.txt
certifi==2020.4.5.1
tensorflow==2.0.0 
chardet==3.0.4
click==7.1.1
cvxpy==1.0.31
cycler==0.10.0
dill==0.3.1.1
ecos==2.0.7.post1
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
multiprocess==0.70.9
numpy==1.18.3
osqp==0.6.1
pandas==0.25.3
pandas-datareader==0.8.1
pyparsing==2.4.7
pyportfolioopt==1.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
scs==2.1.2
seaborn==0.10.0
six==1.14.0
sklearn==0.0
urllib3==1.25.9
Werkzeug==1.0.1
